requirements.txt
kafka-python~=2.2.15
dotenv~=0.9.9
python-dotenv~=1.1.1